+++ src/pages/calendar.html
@@ -21,6 +21,7 @@
     <script type="module" src="/src/components/site-footer.js"></script>
 
     <script type="module" src="/src/calendar.js"></script>
+    <script type="module" src="/src/schedule.js"></script>
     <style>
       /* 
         color-code
@@ -294,9 +295,6 @@
             </template>
           </div>
           <div class="flex justify-center items-end p-1">
-            <button id="prevPage" class="pageBtn hover:cursor-pointer">
-              <i class="fa-solid fa-chevron-left"></i>
-            </button>
             <div id="pageNumDiv">
               <template id="pageNumTemplate">
                 <button class="pageNum text-[#6a994e] hover:cursor-pointer">
@@ -304,9 +302,6 @@
                 </button>
               </template>
             </div>
-            <button id="nextPage" class="pageBtn hover:cursor-pointer">
-              <i class="fa-solid fa-chevron-right"></i>
-            </button>
           </div>
         </div>
       </section>
